requirements.txt
# EPCID - Early Pediatric Critical Illness Detection
# Production Requirements

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.16.0
google-generativeai>=0.3.0
vertexai>=1.38.0

# Async & HTTP
aiohttp>=3.9.0
httpx>=0.25.0
asyncio>=3.4.3

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
# torch>=2.0.0  # Uncomment for production ML
# transformers>=4.30.0  # Uncomment for NLP/LLM features

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# API Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Logging & Monitoring
structlog>=23.2.0
python-json-logger>=2.0.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0

# Security & Authentication
cryptography>=41.0.0
PyJWT[crypto]>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.0
email-validator>=2.0.0

# FHIR
fhir.resources>=7.0.0
fhirclient>=4.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0
ruff>=0.1.0

# Image Processing (optional)
# pillow>=10.0.0
# opencv-python>=4.8.0

# Audio Processing (optional)
# librosa>=0.10.0
# soundfile>=0.12.0
